(feat: ffi) added support for halo2 feature and conditional compilation for all functions (with a stud in case a different proof system is used)
